FormFooter directory reorg
