+++ packages/form-footer/src/index.ts
@@ -1,1 +1,2 @@
-export { default, type FormFooterProps } from './FormFooter';
+export { default } from './FormFooter';
+export { type FormFooterProps } from './FormFooter.types';
